requirements.txt
annotated-types==0.7.0
greenlet==3.0.3
pika==1.3.2
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
typing_extensions==4.12.2
